// src/generated/CreateProjectPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PackageManager } from "./PackageManager";
import type { ProjectLanguage } from "./ProjectLanguage";
import type { ProjectTemplate } from "./ProjectTemplate";
import type { StylingPreset } from "./StylingPreset";

export type CreateProjectPayload = { name: string, directory: string, description: string, template: ProjectTemplate, language: ProjectLanguage, package_manager: PackageManager, dependencies: Array<string>, dev_dependencies: Array<string>, scripts: { [key in string]?: string }, styling: StylingPreset, icon_packs: Array<string>, include_readme: boolean, readme_notes: string, initialize_git: boolean, install_dependencies: boolean, };