// src/generated/DetectedProject.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ScriptEntry } from "./ScriptEntry";

export type DetectedProject = { id: string, path: string, name: string, framework: string, package_manager: string, scripts: Array<ScriptEntry>, has_git: boolean, git_root: string | null, has_env: boolean, env_files: Array<string>, manifests: Array<string>, };