// src/generated/GitHubRepo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type GitHubRepo = { name: string, full_name: string, html_url: string, clone_url: string, ssh_url: string, private: boolean, description: string | null, default_branch: string, updated_at: string, language: string | null, };