// src/generated/ManagedServiceInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ManagedServiceInfo = { service_id: string, cwd: string, cmd: string, pid: number, started_at_ms: number, uptime_ms: number, cpu_usage: number, memory_mb: number, ports: Array<number>, urls: Array<string>, };