// src/generated/ServiceEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ServiceEventKind } from "./ServiceEventKind";

export type ServiceEvent = { service_id: string, kind: ServiceEventKind, message: string, pid: number | null, code: number | null, };